+++ changelog.md
@@ -1,4 +1,3 @@
-# GregTech Modern Utilities version 2.0.0
+# GregTech Modern Utilities version 2.0.1
 
-* Port to GTM 7.0.0
-* Make omni-breaker teleport drops to player on right-click
+* Fix omni-breaker voiding drops on right-click if inventory is full